test: split unit tests into multiple files
